add super class TLNetworkEvent
