+++ relay/blockchain/currency_network_events.py
@@ -1,4 +1,4 @@
-from .events import BlockchainEvent
+from .events import BlockchainEvent, TLNetworkEvent  # NOQA
 
 
 CreditlineRequestEventType = 'CreditlineUpdateRequest'
@@ -9,38 +9,11 @@ BalanceUpdateEventType = 'BalanceUpdate'
 TransferEventType = 'Transfer'
 
 
-class CurrencyNetworkEvent(BlockchainEvent):
+class CurrencyNetworkEvent(TLNetworkEvent):
 
     def __init__(self, web3_event, current_blocknumber, timestamp, user=None):
-        super().__init__(web3_event, current_blocknumber, timestamp)
-        self.user = user
+        super().__init__(web3_event, current_blocknumber, timestamp, from_to_types, user)
         self.network_address = web3_event.get('address')
-
-    @property
-    def from_(self):
-        return self._web3_event.get('args')[from_to_types[self._web3_event.get('event')][0]]
-
-    @property
-    def to(self):
-        return self._web3_event.get('args')[from_to_types[self._web3_event.get('event')][1]]
-
-    @property
-    def direction(self):
-        if self.user is None:
-            return None
-        if self.from_ == self.user:
-            return 'sent'
-        else:
-            return 'received'
-
-    @property
-    def other_party(self):
-        if self.user is None:
-            return None
-        if self.from_ == self.user:
-            return self.to
-        else:
-            return self.from_
 
 
 class ValueEvent(CurrencyNetworkEvent):
